Not removing GET params if autologin was not successful
Unfortunately the "token" GET parameter is used by us to authorize
remp autologin and also by paypal to complete purchase.
Once we tried to autologin user based on GET parameter token and it,
failed, we unset the token, redirected user and paypal wasn't able to
use it for payment confirmation.
Everyone who wasn't already logged in was affected by this bug.

remp/crm#466