Changing HTTPS header to "on" value instead of boolean
This is arbitrary change as any non-empty value should be OK, but Tracy
uses strcasecmp with that value and it causes TypeError if non-string
value is passed inside.

See https://github.com/nette/tracy/blob/master/src/Tracy/Helpers.php#L164
for actual reference of the call causing the issue.