Removing invalid Google column chart bugfix from past, using proper bugfix
The original bugfix altered generated data and forced one day of empty
data to be included in every chart.
This was not correct as only column charts were affected, but data for
all charts were changed.

This bugfix hints date range to Google's chart library so it displays
single-day data correctly.

The original (wrong) bugfix was placed in
7065eac33ad431e23eacd2273f961b74eec45c9f

remp/crm#710