Temporarily enabling and disabling ProductsModule on czech instance
After we moved ProductsModule to the extension, some dependencies from
main config.neon were still using classes from ProductsModule (such as
event handlers in DenniknModule).

Nette tries to initialize them when building DI container and it failed
if products-module extension wasn't enabled (because the requested
classes were not registered in any config.neon file). Because of that,
it was necessary to enable the module and unregister it from application
so it's not visible on the frontend.

Similar issue appeared with injection of dependencies. If the extension
wasn't present, presenters were still being processed by the Injector
early in the process and attempted to use classes from products module.