Fix namespace validation issue with module registration
The issue appeared when namespace was prefixed with multiple
levels (e.g. `Company\Crm\DemoModule`). This is still valid
namespace in the eyes of CRM (and Admin/API ACLs) and should
be allowed.

From now, CRM only validates that module namespace has to end
with *Module word, the rest is up to the implementor.

remp/crm#972