Fix occasional database refetch failed error
The error was caused by improper access to the ActiveRow
data that might have already be invalidated by framework.

If framework decided to refetch the data before accessing
them and the row was already deleted, it would throw
a "database refetch error".

We're now first storing old values from still existing
row record and only after that removing it from the DB.