Add support for measurements
Measurements are meant to replace expensive on-the-fly calculations,
for example number of active subscribers each day/week/month/year.
Measurements generate the values for every day/week/month/year and
can be further grouped by an extra parameter - for example
new registrations by source/channel/etc.

- ScaleFactory for charts now support new "mesaurement" provider,
which uses measurement data as a source of truth. All scales
are available (day/month/week/year).

- Each module needs to register and seed its measurements. Every
measurement is included in the calculation process.

- Command for calculation supports multiple filters, but primary
use for it is:

  1) With no switch to generate all-time data.
  2) With --daily switch to update measurements on day-to-day basis.

- DashboardModule now provides simple display of all available
measurements. Base measurements were also used in dashboard charts as a
reference implementation.

The idea is that each measurement generates "Series" of "Points". In
case measurement supports subgroupping, it also generates
"PointAggregates". These are persistently stored in the database.

remp/crm#1309