Fixed missing host in urls generated by LinkGenerator in CLI
Host is now inferred from CRM_HOST environment variable and passed
to the `http.request` so that `LinkGenerator` can use it.

Since not every environemnt sets $_SERVER['shell'] variable,
we added also $_SERVER['TERM'] variable to determine whether
the application is run from within the CLI or not.

remp/crm#1846