Add support for additional DB (read only) connections
The configuration is done in three steps:

- Add additional database connections to the CRM, so the
application is aware of them.
- Enable decorator allowing use of replicaConfig
in the classes extending Crm\ApplicationModule\Repository.
- Explicitly enable replica connections and DB tables in the config.
Even when all replicas are configured and replicaConfig set,
it's not used unless it's explicitly configured in replicaConfig
setup by allowing specific connections and tables to be used.

All write queries are executed against primary database.

If any of the tables executes write query, repository sets
"write" flag in the replicaManager and any subsequent queries
going to this table are executed against primary database.
Even if they only read the data.

remp/crm#1929