Fix issue CRM admin access due to translation library bug
The bug is causing "en_us_POSIX out of allowed set" error in admin.
It was introduced by symfony/translation starting to return default
php-intl locale instead of nothing. The kdyby/translation was
relying on the previous behavior and is not able to resolve the
translation correctly after the change.

Commit causing the issue:
https://github.com/symfony/translation/blob/db0ba1e85280d8ff11e38d53c70f8814d4d740f5/Translator.php#L175

kdyby/translation relying on empty locale:
https://github.com/Kdyby/Translation/blob/4c4a02a22922b8d69961bb4cd17c104c451622a4/src/Translator.php#L352

remp/crm#1998