Fix deprecation notice caused by autowiring asking for Presenter::getContext
The issue was caused by autowiring extension (kdyby/autowired) which asks for
Presenter::getContext if it doesn't have component locator initialized from
the beginning by the DI.

This change forces the component locator to be injected during DI, so it
doesn't have to be requested on-the-fly. More info in the related issue.

remp/crm#2214