Fix session regeneration on each request for non-autologged admins
When admin account tried to use AccessTokenAuthenticator to
autologin into CRM based on the n_token cookie, it correctly
failed because of the security features.

After the update to Nette 3, login attempts reset session ID.

This caused that on each request, authenticator attempted to
log user in, each time it failed and each time it generated
new session ID. That made session for these kind of users
unusable (e.g. they weren't able to add products to session-
based cart).

Authentication is now moved to the separate event handler out
of the FrontendPresenter (which was our intention anyway)
and is only executed once. After the failed attempt the session
is flagged and never attempted again.

Co-Authored-By: Dominik Skerhak <dominik.skerhak@gmail.com>

remp/crm#2328