AccessTokenAuthenticator - store access token hash instead of flag of disabled autologin, minor refactoring
Changed way how `AccessTokenAuthenticator` marks user token that was not allowed to login. 
Previously, only a flag (SESSION_AUTH_DISABLED) was stored in the session. Now, hash of a token is stored.
This fixes a problem when user logs out, actual token changes, but user cannot log in because the flag is still present in the session storage. Now this is correctly handled since hashes of tokens are compared