Revert second-layer referer resolving
The feature was reverted back as it was still actively used in the sales
funnels.

Funnels submit their contents to the /submit URL which doesn't contain
referer query parameter. The referer header parameter was the actual
URL of the funnel. And since at no point funnels actively add referer
to the GET or POST parameters of the submitted data, we decided
to rollback the feature for now.

remp/crm#2662