Fix discrepancies in localized date formatting
We need to match against one of the two possible values, because of the
change in the ICU system library 72.1.

Since it's not possible to determine current version, and there might or
might not be a non-breaking space between the time and AM/PM, we just
test this against one of the two values.

Official php-fpm docker image uses ICU v67.1, but the php-apline uses
72.1 and this behaved unexpectedly.