Changed rendering options of single-value admin charts
Y-axis now doesn't always start from zero in favor of readability. If
the difference between extremes of the chart is less than 30% of
maximum's value, the Y-axis will not start from zero anymore.

This is to improve readability of high-value charts which change, but
the change is not that visible.

remp/crm#2788